Split SummaryGenerator into its own file with SummarizerHeuristics
